*Fjerner Unleash kall (ettersom backend ikke eksisterer for øyeblikket) *Bugfixes
