+++ src/providers/Store.ts
@@ -90,7 +90,7 @@ export const reducer = (state: Store, action: Action) => {
     case "SETT_FEATURE_TOGGLES":
       return {
         ...state,
-        unleashFeatures: Unleash.getValidFeatures(action.payload as Features)
+        unleashFeatures: Unleash.getValidFeatureToggles(action.payload as Features)
       };
     default:
       return state;
